% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Growth.R
\name{growth_objective}
\alias{growth_objective}
\title{Create growth objective function with data baked in}
\usage{
growth_objective(datain)
}
\arguments{
\item{datain}{List containing model data}
}
\value{
Function that takes parameters and returns negative log-likelihood
}
\description{
Create growth objective function with data baked in
}
