% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Growth.R
\name{make_growmod_obj}
\alias{make_growmod_obj}
\title{Create RTMB objective from growmod with proper environment}
\usage{
make_growmod_obj(pin, datain = NULL, map = list(), random = NULL)
}
\arguments{
\item{pin}{Parameter list}

\item{datain}{Data list (must be in calling environment)}

\item{map}{Optional parameter map}

\item{random}{Optional random effects}
}
\description{
This function copies growmod to the global environment to avoid
namespace issues with RTMB's automatic differentiation.
}
